+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+# v0.0.9
+
+**Breaking change:**
+- Updated return response for upload(file, signedDetails,callback,chunkSize,concurrency) API
+- On success, response.data will now return an instance UploadResponse instead String.
+
 # v0.0.8
 
 - backward compatibility to gcs
 - abstracting different file upload methods to single method
-
 
 # v0.0.7